requirements.txt
asgiref==3.7.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2
djangorestframework==3.14.0
grpcio==1.54.2
grpcio-tools==1.54.2
grpclib==0.4.4
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
multidict==6.0.4
protobuf==4.23.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
